requirements.txt
Django==4.2
django-4-jet==1.0.9
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Jinja2==3.1.4
pillow==11.0.0
requests==2.32.3
whitenoise==6.8.2